fix: исправлена обработка запросов

+++ client/src/indexedDB/api_indexedDB/processPendingRequests.js
@@ -11,31 +11,36 @@ export const processPendingRequests = async () => {
   const tx = db.transaction("pendingRequests", "readonly");
   const pendingRequestsStore = tx.objectStore("pendingRequests");
   const request = await pendingRequestsStore.getAll();
-  request.onsuccess = (event) => {
-    const allRequests = event.target.result;
-    allRequests.forEach(({ type, data }) => {
-      switch (type) {
-        case "create":
-          addNoteToMongo(data);
-          break;
-        case "delete":
-          deleteNoteFromMongo(data.id);
-          break;
-        case "edit":
-          editNoteFromMongo(data.id, data);
-          break;
-        default:
-          break;
+
+  return new Promise((resolve) => {
+    request.onsuccess = async (event) => {
+      const allRequests = event.target.result;
+      if (allRequests.length === 0) {
+        // Если отложенных запросов нет, просто закрываем базу данных
+        db.close();
+        resolve();
+        return; // Прерываем выполнение функции
       }
-    });
-  };
 
-  // Очистка отложенных запросов после их обработки
-  const clearTx = db.transaction("pendingRequests", "readwrite");
-  const clearStore = clearTx.objectStore("pendingRequests");
-  clearStore.clear();
+      for (const { type, data } of allRequests) {
+        if (type === "create") {
+          await addNoteToMongo(data);
+        } else if (type === "delete") {
+          await deleteNoteFromMongo(data.id);
+        } else if (type === "edit") {
+          await editNoteFromMongo(data._id, data);
+        }
+      }
 
-  clearTx.oncomplete = () => {
-    db.close();
-  };
+      // Очистка отложенных запросов после их обработки
+      const clearTx = db.transaction("pendingRequests", "readwrite");
+      const clearStore = clearTx.objectStore("pendingRequests");
+      clearStore.clear();
+
+      clearTx.oncomplete = () => {
+        db.close();
+        resolve();
+      };
+    };
+  });
 };
